Respect --test-threads options and clean up

+++ src/args.rs
@@ -93,7 +93,7 @@ pub struct Arguments {
         long = "--test-threads",
         help = "Number of threads used for running tests in parallel",
     )]
-    pub num_threads: Option<u32>,
+    pub num_threads: Option<usize>,
 
     /// Path of the logfile. If specified, everything will be written into the
     /// file instead of stdout.
